requirements.txt
# API Clients
openai>=1.3.0
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Type hints support
typing-extensions>=4.8.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Optional: Development tools
black>=23.11.0  # for code formatting
pytest>=7.4.3   # for testing
pytest-asyncio>=0.23.2  # for testing async code
